refactored and updated colors
